Move overload resolution into Binding.OverloadResolution namespace
Extract the overload-resolution cluster into a dedicated sub-namespace,
GSharp.Core.CodeAnalysis.Binding.OverloadResolution, following the
Lowering.Async / Lowering.Iterators directory-per-namespace convention:

- OverloadResolver.cs and its five partials (Arguments, CallBinding,
  Candidates, Constructors, Invocations) - the binder collaborator that
  resolves user-function, extension, constructor, and invocation
  candidates.
- ClrOverloadResolution.cs - the pure CLR MethodBase-candidate
  resolution algorithm and its argument sentinels.

Each moved file changes only its namespace declaration; consumers gain
a single using directive. Conversion machinery (Conversion,
ConversionClassifier) and operator resolution (ClrOperatorResolution,
EnumOperatorTable) stay in Binding: they are general binding
infrastructure consumed by pattern/statement binding and the emitter,
not overload-resolution internals.

Part of #3163. Closes #1365.
