+++ src/Core/CodeAnalysis/Binding/NumericWideningLattice.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Collections.Generic;
+using GSharp.Core.CodeAnalysis.Binding.OverloadResolution;
 
 namespace GSharp.Core.CodeAnalysis.Binding;
