+++ src/Core/CodeAnalysis/Binding/PatternBinder.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Immutable;
 using System.Numerics;
 using System.Reflection;
+using GSharp.Core.CodeAnalysis.Binding.OverloadResolution;
 using GSharp.Core.CodeAnalysis.Symbols;
 using GSharp.Core.CodeAnalysis.Syntax;
