+++ src/Core/CodeAnalysis/Symbols/ReferenceResolver.cs
@@ -15,6 +15,7 @@ using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
 using System.Runtime.CompilerServices;
 using GSharp.Core.CodeAnalysis.Binding;
+using GSharp.Core.CodeAnalysis.Binding.OverloadResolution;
 
 namespace GSharp.Core.CodeAnalysis.Symbols;
 
@@ -658,8 +659,8 @@ public sealed class ReferenceResolver : IDisposable
         // without being rebound. Fail with an actionable internal-compiler-error
         // that names the bug instead of the opaque MetadataLoadContext projection
         // failure it would otherwise cause.
-        if (ReferenceEquals(hostType, GSharp.Core.CodeAnalysis.Binding.ClrOverloadResolution.InlineOutVarArgumentType)
-            || ReferenceEquals(hostType, GSharp.Core.CodeAnalysis.Binding.ClrOverloadResolution.DefaultLiteralArgumentType))
+        if (ReferenceEquals(hostType, ClrOverloadResolution.InlineOutVarArgumentType)
+            || ReferenceEquals(hostType, ClrOverloadResolution.DefaultLiteralArgumentType))
         {
             throw new InvalidOperationException(
                 $"Internal compiler error: the overload-resolution sentinel '{hostType.FullName}' escaped the " +
