+++ test/Core.Tests/CodeAnalysis/Binding/Issue2516SliceCovarianceBindingTests.cs
@@ -5,6 +5,7 @@
 using System;
 using System.Collections.Generic;
 using GSharp.Core.CodeAnalysis.Binding;
+using GSharp.Core.CodeAnalysis.Binding.OverloadResolution;
 using GSharp.Core.CodeAnalysis.Symbols;
 using Xunit;
